+++ smart-home-api/src/main/java/de/markostreich/smarthome/leddeviceapi/model/dto/LedStripeObjectDto.java
@@ -3,8 +3,8 @@ package de.markostreich.smarthome.leddeviceapi.model.dto;
 public record LedStripeObjectDto(
 		String name,
 		String mode,
-		byte red,
-		byte green,
-		byte blue,
-		byte brightness,
+		int red,
+		int green,
+		int blue,
+		int brightness,
 		String deviceName) {}
